% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny app}
\usage{
launch_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyAppDir()}}.}
}
\value{
No return value; called for side effects.
}
\description{
Launches the Shiny application bundled in this package under \verb{inst/app/}.
}
\examples{
if (interactive()) {
  myappkg::launch_app()
}
}
