% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-doc.R
\docType{data}
\name{clean_dataset}
\alias{clean_dataset}
\title{cleaned data for the shiny app
Description for the variables here.}
\format{
A data frame with 100 observations and 3 variables

\describe{
    \item{R0}{the spread capability of the virus,ranging from 1 to 10}
    \item{VE}{the vaccine efficiency, ranging from 0.0 to 0.9}
}
}
\source{
original course dataset, cleaned by data-raw/clean_dataset.R
}
\usage{
clean_dataset
}
\description{
cleaned data for the shiny app
Description for the variables here.
}
\keyword{datasets}
